Expose input\windows Config on all OSes
